Seperate files in context package(genericclioptions)
