--- pkg/odo/genericclioptions/context_flags.go
@@ -1,14 +1,0 @@
-package genericclioptions
-
-const (
-	// ProjectFlagName is the name of the flag allowing a user to specify which project to operate on
-	ProjectFlagName = "project"
-	// ApplicationFlagName is the name of the flag allowing a user to specify which application to operate on
-	ApplicationFlagName = "app"
-	// ComponentFlagName is the name of the flag allowing a user to specify which component to operate on
-	ComponentFlagName = "component"
-	// OutputFlagName is the name of the flag allowing user to specify output format
-	OutputFlagName = "output"
-	// ContextFlagName is the name of the flag allowing a user to specify the location of the component settings
-	ContextFlagName = "context"
-)
